Save work

+++ mysql/deleteAttendanceCategory.ts
@@ -1,7 +1,6 @@
 import pool from "./pool";
 import { OkPacket, RowDataPacket } from "mysql2";
 
-
 async function deleteAttendanceCategory(categoryUID: string): Promise<{ querySuccess: boolean, error?: any}> {
     const promisePool = pool.promise();
 
@@ -10,10 +9,15 @@ async function deleteAttendanceCategory(categoryUID: string): Promise<{ querySuc
         .then(connection => {
             connection.beginTransaction()
             .then(async () => {
-                //here check if the category has entry
+                //check if the category has entry
                 //if true cancel delete by throwing error
                 //else continue detetion
-                //code soon
+                const hasEntries = (await connection.query(`
+                SELECT COUNT(*) AS totalEntries
+                FROM attendance_entries AS ae
+                WHERE ae.category_uid = ?`, [categoryUID]) as RowDataPacket[][])[0][0].totalEntries;
+
+                if(hasEntries) throw "Cannot delete category that has entries";
 
                 //Delete attendance category query
                 connection.query(`DELETE FROM attendance_categories WHERE uid = ?`, [categoryUID]);
